services/requirements.txt
# QuantEmbrace - Python Dependencies
# ====================================

# Broker SDKs
kiteconnect>=5.0.0            # Zerodha Kite Connect API (NSE India)
alpaca-trade-api>=3.0.0       # Alpaca Trading API (US Equities)

# AWS SDK
boto3>=1.34.0                 # AWS services (S3, DynamoDB, SQS, SNS)
aiobotocore>=2.12.0           # Async AWS SDK

# Web Framework (AI Engine)
fastapi>=0.110.0              # HTTP API for AI engine
uvicorn[standard]>=0.29.0     # ASGI server

# Data Models & Validation
pydantic>=2.6.0               # Data validation and models
pydantic-settings>=2.2.0      # Settings management

# Data & ML
numpy>=1.26.0                 # Numerical computing
pandas>=2.2.0                 # Data manipulation
scikit-learn>=1.4.0           # ML models
joblib>=1.3.0                 # Model serialization

# Async & Networking
aiohttp>=3.9.0                # Async HTTP client
websockets>=12.0              # WebSocket client for market data

# Logging & Monitoring
structlog>=24.1.0             # Structured logging

# Testing
pytest>=8.1.0                 # Test framework
pytest-asyncio>=0.23.0        # Async test support
pytest-cov>=4.1.0             # Coverage reporting
hypothesis>=6.98.0            # Property-based testing
responses>=0.25.0             # HTTP mocking
moto>=5.0.0                   # AWS mocking (S3, DynamoDB, SQS)

# Code Quality
black>=24.3.0                 # Code formatter
ruff>=0.3.0                   # Linter
isort>=5.13.0                 # Import sorting
mypy>=1.9.0                   # Type checking
